# Conflicts:
#	build/tmp/compileJava/previous-compilation-data.bin
#	src/main/java/com/sparta/newsfeed/newsfeed/service/NewsfeedServiceImpl.java

+++ src/main/java/com/sparta/newsfeed/user/service/UserService.java
@@ -3,10 +3,7 @@ package com.sparta.newsfeed.user.service;
 import com.sparta.newsfeed.auth.service.AuthService;
 import com.sparta.newsfeed.config.PasswordEncoder;
 import com.sparta.newsfeed.newsfeed.repository.NewsfeedRepository;
-import com.sparta.newsfeed.user.dto.SignUpRequestDto;
-import com.sparta.newsfeed.user.dto.SignUpResponseDto;
-import com.sparta.newsfeed.user.dto.UserRequestDto;
-import com.sparta.newsfeed.user.dto.UserResponseDto;
+import com.sparta.newsfeed.user.dto.*;
 import com.sparta.newsfeed.user.entity.User;
 import com.sparta.newsfeed.user.repository.UserRepository;
 import jakarta.servlet.http.HttpSession;
@@ -15,8 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-
-import java.util.UUID;
 
 @Transactional
 @Service
@@ -31,13 +26,11 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
 
     public SignUpResponseDto createUser(SignUpRequestDto dto) {
-        System.out.println("iddddddd");
         if (userRepository.existsUserByEmail(dto.getEmail())) {
             // 해당 이메일의 사용자가 탈퇴했는지 여부는 알 수 없음
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "가입이 불가능한 이메일입니다.");
         }
         User user = new User(dto.getName(), dto.getEmail(), passwordEncoder.encode(dto.getPassword()));
-        System.out.println("iddddddd22222");
         User savedUser = userRepository.save(user);
 
         return new SignUpResponseDto(savedUser.getName(), savedUser.getEmail());
@@ -50,14 +43,10 @@ public class UserService {
     }
 
     public UserResponseDto updateUser(Long id, UserRequestDto requestDto, HttpSession session){
+
         //TODO: 에러 처리
         User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
-
         String sessionEmail = authService.getSessionEmail(session);
-
-        if (requestDto.getOldPassword().equals(requestDto.getNewPassword())) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You can not change with same password.");
-        }
 
         if (!foundUser.getEmail().equals(sessionEmail)) {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You can not update other user's information.");
@@ -68,22 +57,34 @@ public class UserService {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is wrong.");
         }
 
+        if (requestDto.getOldPassword().equals(requestDto.getNewPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You can not change with same password.");
+        }
+
         foundUser.partialUpdate(
                 requestDto.getName(),
-                requestDto.getEmail(),
                 passwordEncoder.encode(requestDto.getNewPassword())
         );
 
         return new UserResponseDto(id, requestDto.getEmail(), requestDto.getName());
     }
 
+    public void deleteUser(Long id, UserDeleteRequestDto requestDto , HttpSession session){
+        User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
+        String email = authService.getSessionEmail(session);
 
-    public void deleteUser(Long id, HttpSession session){
+        if (!email.equals(foundUser.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "You can not delete other user's account.");
+        }
+
+        //TODO:겹치는 코드 , 메서드 추출 ?
+        if (!passwordEncoder.matches(requestDto.getPassword(), foundUser.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is wrong.");
+        }
+
         newsfeedRepository.deleteNewsfeedsByUserId(id);
         //TODO: 에러 처리
-        User foundUser = userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id));
         foundUser.updateSoftDelete();
-
         authService.logout(session);
     }
